(WIP) Using GeoIP2 instead of Pygeoip, and add DEFAULT_COUNTRY
pygeoip works only on IPv4 or IPv6 with a single database, while in a dual-stack configuration, both IPv4 and IPv6 addresses may occur in the tracking log. In this case, some logs may not be correctly parsed and the location reports will be wrong.

Another issue here is that currently all IP addresses not in the database will be marked as 'UNKNOWN'. However, in some openedx deployments, especially those inside an institute or university, users inside the institute or university may use private adress to access openedx, which will generate a lot of 'UNKNOWN' results with the current code. This PR uses DEFAULT_COUNTRY rather than UNKNOWN_COUNTRY for such private addresses.