Use a retry library to retry calls to the HistoryServer API.
It seems like the HistoryServer can get a little choked up on bigger
jobs, or by calling it too quickly after the jobs finish.

Now we have a little exponential backoff going on.  We'll try a total of
6 times to collect the metrics, with an exponential backoff of
2^n*1000ms (n is number of retries) with a maximum of 10000ms delay.

So... 0 seconds delay, then 1s, then 4s, then 10s, and 10s until the
retries are up.