Trap exceptions that occur while collecting metrics.
In all cases, we do not want a metrics collection issue to actually fail
a workflow in the eyes on Jenkins.  We specifically look for traceback
output to signal us to a job that had an error, but this check also
catches errors from the metrics collection.

Now, we'll catch exceptions and simply log their message, rather than
dumping out the traceback.  This should be sufficient for help guide
further debugging without triggering any of the checks that we use to
introspect the status of the workflow itself.