Allow multiple secure configurations to be loaded.
This change allows a caller to specify multiple configuration files in
the secure config repo to be loaded, which enables the caller to specify
a "base" configuration, containing special values, which can be
interpolated into subsequent configuration files that are loaded.

These configuration paths are now passed on from remote-task itself to
launch-task when invoked on an EMR cluster.