Weave through paths in a more consistent fashion.
We're now properly including 'uid', which is hard-coded to 'automation'
in production, to the paths.  This should make the code more resilient
to cases where no remote name is specified.

We've also moved the constants back into remote.py, given that they
weren't necessary in local.py and thus did not need to be shared.