When writing to S3, don't validate the bucket.
If Boto is passed "validate=True", it will require an additional
permission to be present when asked to list all of the keys in the
bucket.  We want to minimize the set of required permissions so we get a
reference to the bucket without validating that it exists.  It should
only require PutObject and PutObjectAcl permissions in order to write to
the target bucket.

This is particularly the case with answer-distribution, which writes out
to a bucket that is used by LMS for course downloads.