Add new f_orderline_transactions table to finance-report workflow (#736)
The main differences of f_orderline_transactions versus f_orderitem_transactions is the addition of four new columns and the removal of one. The additions are "order_product_course_id" (the value of course_id stored in the product, rather than as an attribute value), "order_lms_user_id" (as stored in ecommerce_user), "order_partner_sku" (as requested long ago), and "order_status". Also, the set of orders that are included in the table are slightly expanded in two ways: one is the inclusion of orderlines for products that don't have "parent" products (so these are standalone and missing products), and second is the inclusion of orders that don't have a status of "Complete". (If you don't want them, you can filter them using the new "order_status" column.)

Also adds loads of the orders tables that go into the f_orderline_transactions tables, for better auditing (as two separate tables: "full_otto_orders" and "full_shoppingcart_orders").

By piggy-backing on the finance-report workflow, no new DSL job needs to be created, or coordinated with the finance-report job.

* Not all INNER JOINS are converted to LEFT JOINs.  Orders and users change often enough that it matters that they are pulled at different times, and they require INNER JOINs to maintain consistency.

* Split orders into Shoppingcart and Otto, so they
can be loaded separately into the warehouse (as an artifact,
at the very least).