Pin pip, setuptools, wheel to fix opaque keys issue
Currently EMR is failing to package up .egg-info since we un-pinned pip and friends. This means that stevedore plugins that rely on that to find plugins can't find them, which breaks parsing of new-style course keys and other things. This change should get us back to a time when all of those things played nicely together.