docs: Fix `make docs` to actually build.
A re-arrangement of the requirements files resulted in us needing to
install postgres in order to build docs which requires the postgres
binary to be install on the base system.  This is not possible on the
read-the-docs machines so I updated the docs.in file to explicitly
require anything it needs to be able to generate docs.